+++ DavidCCompilerStudy/dynarray.h
@@ -3,11 +3,11 @@
 
 typedef struct DynamicArray {
 	void** array;         // array of void *s
-	uint16_t capacity;        // max 65,536
-	uint16_t size; // max 65,536
+	uint32_t capacity;        // max 65,536
+	uint32_t size; // max 65,536
 } DynamicArray;
 
-// returns a dynamic array "INITIAL_DYNARRAY_CAPACITY (16) big
+// returns a dynamic array "INITIAL_DYNARRAY_CAPACITY (4) big
 DynamicArray* dd_makeDynamicArray();
 
 // pushes a new item to the end of the dynamic array
@@ -20,10 +20,10 @@ void* dd_pop(DynamicArray* dynarray);
 void* dd_free(DynamicArray* dynarray);
 
 // get void * indexed at "index"
-void* dd_get(DynamicArray* dynarray, uint16_t index);
+void* dd_get(DynamicArray* dynarray, uint32_t index);
 
 // set void * indexed at "index"
-void dd_set(DynamicArray* dynarray, uint16_t index, void* newItem);
+void dd_set(DynamicArray* dynarray, uint32_t index, void* newItem);
 
 // resize the array to the specified capacity
-void dd_resize(DynamicArray* dynarray, uint16_t capacity);
+void dd_resize(DynamicArray* dynarray, uint32_t capacity);
